+++ SlipeServer.Example/LuaControllersExampleLogic.cs
@@ -1,4 +1,7 @@
 ﻿using SlipeServer.LuaControllers.Commands;
+using SlipeServer.Server.ElementCollections;
+using SlipeServer.Server.Elements;
+using SlipeServer.Server.Services;
 
 namespace SlipeServer.Example;
 
@@ -9,7 +12,10 @@ public class SampleClass
 
 public class LuaControllersExampleLogic
 {
-    public LuaControllersExampleLogic(LuaControllerArgumentsMapper mapper)
+    private readonly IElementCollection elementCollection;
+    private readonly ChatBox chatBox;
+
+    public LuaControllersExampleLogic(LuaControllerArgumentsMapper mapper, IElementCollection elementCollection, ChatBox chatBox)
     {
         mapper.DefineMap<SampleClass>(arg =>
         {
@@ -18,5 +24,24 @@ public class LuaControllersExampleLogic
                 Number = int.Parse(arg)
             };
         });
+        mapper.DefineMap<Player>(arg =>
+        {
+            return elementCollection.GetByType<Player>().Where(x => x.Name.Contains(arg)).FirstOrDefault();
+        });
+
+        mapper.ArgumentErrorOccurred += HandleArgumentErrorOccurred;
+        this.elementCollection = elementCollection;
+        this.chatBox = chatBox;
+    }
+
+    private void HandleArgumentErrorOccurred(Player player, LuaControllerArgumentException ex)
+    {
+        if(ex.InnerException is ArgumentOutOfRangeException)
+        {
+            this.chatBox.OutputTo(player, "Too many or too few arguments");
+        } else
+        {
+            this.chatBox.OutputTo(player, "Error while executing command");
+        }
     }
 }
